+++ NEWS.md
@@ -1,2 +1,8 @@
+0.2.0
+* Functions attempt to auto-derive the iOS backup location on macOS and legacy Windows operating systems.
+* Uses the `{fs}` pacakge for all file/directory operations
+* Provides a `list_ios_backups()` function so you can see all the iOS backups available
+* `src_mowerplus()` defaults to using the most recent iOS backup (if it exists) if unspecified
+
 0.1.0 
 * Initial release